# doc/api.yaml
openapi: 3.1.0
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto!

    You can upload your photos directly from your PC, and they will be visible to everyone following you.
  version: 1.0.0
components:
  schemas:
    User:
      title: User
      description: This object represents a single user
      type: object
      properties:
        userId:
          type: integer
          description: Unique identifier for the user
          example: 1234
          readOnly: true
          minimum: 0
        username:
          description: The username displayed to the others
          type: string
          example: CoolUsername42
          minLength: 3
          maxLength: 32
          pattern: '^[A-Za-z0-9_\-]{3,32}$'
      required: [ "userId", "username" ]
    Profile:
      title: Profile
      description: This object represents the profile data of a user
      type: object
      properties:
        userId: { $ref: '#/components/schemas/User/properties/userId' }
        username: { $ref: '#/components/schemas/User/properties/username' }
        photoCount:
          description: The count of how many photos this user uploaded
          type: integer
          example: 42
          minimum: 0
        followersCount:
          description: The amount of users following this user
          type: integer
          example: 42
          minimum: 0
        followingCount:
          description: The amount of users followed by this user
          type: integer
          example: 42
          minimum: 0
        bannedCount:
          description: The amount of users banned by this user
          type: integer
          example: 42
          minimum: 0
      required: [ "userId", "username", "photoCount", "followersCount", "followingCount", "bannedCount" ]
    Comment:
      title: Comment
      description: This object represents a single comment bound to a Photo
      type: object
      properties:
        commentId:
          type: integer
          description: Unique identifier for the comment
          example: 1234
          minimum: 0
          readOnly: true
        ownerId: { $ref: '#/components/schemas/User/properties/userId' }
        ownerUsername: { $ref: '#/components/schemas/User/properties/username' }
        content:
          type: string
          description: This is the content of the comment written by the owner
          example: "That's a nice image!"
          pattern: '^[\p{L}\p{N}\p{M}\p{P}\p{S} ]{1,256}$'
          minLength: 1
          maxLength: 256
      required: [ "commentId", "ownerId", "content" ]
...
    Error:
      title: Error
      description: |-
        This object represents a single error to be returned for certain requests
      type: object
      properties:
        code:
          type: integer
          description: An HTTP error code
          example: 404
        message:
          type: string
          description: A brief message describing the error code
          example: The specified resource was not found
      required: [ "code", "message" ]
  parameters:
    usernameParam:
      name: username
      in: path
      description: Username of the user
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/User/properties/username' }
    targetUsernameParam:
      name: targetUsername
      in: path
      description: Username of the target
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/User/properties/username' }
    photoIdParam:
      name: photoId
      in: path
      description: ID of the photo
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Photo/properties/photoId' }
...
    commentIdParam:
      name: commentId
      in: path
      description: ID of the comment
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Comment/properties/commentId' }
  responses:
...
    NoContent:
      description: The request was successful, but there is no content to return.
    # 400
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 401
    Unauthorized:
      description: Invalid or expired authentication credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 404
    NotFound:
      description: The specified resource was not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 409
...
    UnprocessableEntity:
      description: The request was well-formed but was unable to be followed due to semantic errors.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    # 500
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: [ ]
...
paths:
...
  /users/{username}/profile:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    get:
      tags: [ "User" ]
      operationId: getUserProfile
      summary: Retrieve the information of a user
      description: |-
        Given a user's username, retrieve all the public info available.
      responses:
        "200":
          description: Successfully retrieved profile data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/bans:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    post:
      tags: [ "Ban" ]
      operationId: banUser
      summary: Ban a user
      description: Given a user's username, ban them.
      requestBody:
        description: The username of the user to ban
        required: true
        content:
          application/json:
            schema:
              type: object
              description: username schema
              properties:
                username: { $ref: '#/components/schemas/User/properties/username' }
      responses:
        "201":
          description: User banned successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/bans/count:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    get:
      tags: [ "Ban" ]
      operationId: getBansCount
      summary: Get the count of users banned by the current user
      description: Get the count of users banned by the current user
      responses:
        "200":
          description: Successfully retrieved the count of banned users
          content:
            application/json:
              schema:
                type: integer
                description: count of banned users
                example: 42
                minimum: 0
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/bans/list:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
    get:
      tags: [ "Ban" ]
      operationId: getBansList
      summary: Get the list of users banned by the current user
      description: Get the list of users banned by the current user
      responses:
        "200":
          description: Successfully retrieved the list of banned users
          content:
            application/json:
              schema:
                type: array
                description: list of usernames
                items: { $ref: '#/components/schemas/User/properties/username' }
                uniqueItems: true
                minItems: 0
                maxItems: 99999
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/bans/list/{targetUsername}:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
      - $ref: '#/components/parameters/targetUsernameParam'
    get:
      tags: [ "Ban" ]
      operationId: getBanStatus
      summary: Check if a user is banned
      description: |-
        Given a user's username, check if they are banned.
        In this case, the target user is the 1st parameter.
      responses:
        "200":
          description: User is banned
          content:
            application/json:
              schema:
                type: object
                properties:
                  isFollowing:
                    type: boolean
                    description: Whether the user has banned the target user
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/bans/{targetUsername}:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
      - $ref: '#/components/parameters/targetUsernameParam'
    delete:
      tags: [ "Ban" ]
      operationId: unbanUser
      summary: Unban a user
      description: Given a user's username, unban them.
      responses:
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...
  /users/{username}/photos/{photoId}/comments:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
      - $ref: '#/components/parameters/photoIdParam'
    get:
      tags: [ "Comment" ]
      operationId: getPhotoComments
      summary: Retrieve the comments of a photo
      description: |-
        Given the id of a photo from a user, retrieve all its comments.
      responses:
        "200":
          description: Successfully retrieved comments
          content:
            application/json:
              schema:
                type: array
                description: list of comments
                items: { $ref: '#/components/schemas/Comment' }
                uniqueItems: true
                minItems: 0
                maxItems: 99999
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
    post:
      tags: [ "Comment" ]
      operationId: commentPhoto
      summary: Add a comment to a photo
      description: |-
        Given the id of a photo from a user (defined by username), add a comment to it.
        Multiple comments can be created by the same user.
      requestBody:
        description: The comment content
        required: true
        content:
          application/json:
            schema:
              type: object
              description: comment content schema
              properties:
                content: { $ref: '#/components/schemas/Comment/properties/content' }
      responses:
        "201":
          description: Comment added successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comment' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
  /users/{username}/photos/{photoId}/comments/{commentId}:
    parameters:
      - $ref: '#/components/parameters/usernameParam'
      - $ref: '#/components/parameters/photoIdParam'
      - $ref: '#/components/parameters/commentIdParam'
    delete:
      tags: [ "Comment" ]
      operationId: uncommentPhoto
      summary: Remove a comment from a photo
      description: |-
        Given the id of a comment from a photo (defined by id as well), delete it.
        Only the comment owner can delete it.
      responses:
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }